bare minimum almost done
Main algorithm for comparison works accurately for both gapped and
ungapped. Domains can now be specified in either a range ex: [1 - 22],
specific points ex:[1, 2, 3, 10], and a mix of ranges and specific
points ex: [1, 4, [6-10], 11, 13], and does it automatically.

Custom domains and the model name are also now retrieved the
Domains.txt. The regular domains are retrieved from another websource
and runs a selenium script to grab the domains. The sequence for both
the model and templates are now gathered from the .fasta file.

setup.py runs a serious of unix command line commands to download
selenium, download the chrome web driver, and place the web driver in
the proper system location to ease installation for those unfamiliar
with command line.